requirements.txt
numpy
pandas
requests
matplotlib
einops
tqdm
pytorch-lightning
scikit-learn
opencv-python
neptune-client
Pillow